style: organizes the global constants
There where to many colors into the constants file, so now they have their own file, and you can see clearly what are the colors and what are the other constants.
Some colors names where updated too, to see clearly the difference between all the available colors.
