+++ src/colors.js
@@ -1,0 +1,31 @@
+export const principalBlue = "#0E4B90";
+export const transparentPrincipalBlue = "rgb(14, 75, 144, 0.5)";
+export const principalLightBlue = "#84C8EF";
+export const mediumBlue = "#0095d4";
+
+export const disabledGray = "#bbc1c6";
+
+export const lightOrange = "#FFE8DF";
+
+export const gradientPalettes = {
+  1: [principalBlue],
+  2: [principalBlue, principalLightBlue],
+  3: [
+    principalBlue,
+    "#498AC0",
+    principalLightBlue
+  ],
+  4: [
+    principalBlue,
+    "#3575B0",
+    "#5D9ECF",
+    principalLightBlue
+  ],
+  5: [
+    principalBlue,
+    "#2C6AA8",
+    "#498AC0",
+    "#67A9D7",
+    principalLightBlue
+  ],
+}
