+++ src/constants.js
@@ -135,31 +135,3 @@ export const tallyNames = [
   ...firstMajorityTallyNames,
   ...preferentialRankingTallyNames,
 ];
-
-export const principalBlue = "#0E4B90";
-export const transparentPrincipalBlue = "rgb(14, 75, 144, 0.5)";
-
-export const principalLightBlue = "#84C8EF";
-
-export const gradientPalettes = {
-  1: [principalBlue],
-  2: [principalBlue, principalLightBlue],
-  3: [
-    principalBlue,
-    "#498AC0",
-    principalLightBlue
-  ],
-  4: [
-    principalBlue,
-    "#3575B0",
-    "#5D9ECF",
-    principalLightBlue
-  ],
-  5: [
-    principalBlue,
-    "#2C6AA8",
-    "#498AC0",
-    "#67A9D7",
-    principalLightBlue
-  ],
-}
